Apprentices are not patrol leaders unless patrol is only apprentices
ozzie ☁#2473 in #dev-bug-reports pointed out that apprentices could be chosen as patrol leaders even if they were not the most experienced cat, so now when creating a patrol, each cat will be checked if their status is apprentice or not. If not, they are added to list possible_patrol_leaders and a patrol leader is picked from that. If the patrol only contains apprentices, it will pick a random apprentice to be the leader