Heterochromia go!
Heterochromia functions by having sheets eyes2, eyesextra2, and eyesparalyzed2 which contain the right eye only. If a cat is heterochromic, a second eye color (eye_colour2) is generated that is guaranteed to be different from the first. This allows for any combination of eye colors with very little extra work for artists.

Needs a bit of cleanup, but everything works.