Added a whole buncha "nulls" to the forest hunting .json
Trying to fix a bug where the game won't give freshkill with these patrols so I added 'null' to every line that should have an outcome according to the patrol writing bible so it HOPEFULLY returns and gives the correct freshkill. Also added "medium_prey" to some patrols that had "small_prey" for balancing reasons.