emptied both med app and elder .json
there's 9 relationship to other jsons, and I've done 4. About 2/5ths. could be better, could be worse. Lot of repeated events in the next 3 I plan to tackle next, which are warrior, deputy, and leader. Medicine and app have the most unique events, but should be manageable after I remove the events that are also present elsewhere.