No more triple letter names
With help from Lixxis and Key, we were able to code in a check for triple letter names that occur when a prefix (the first half) has two letters at the end of it (Moss, Grass, Small, etc.) and a suffix (the last half) with the same letter get joined (Mossscratch, Grassseed, Smallleaf, ect.) preventing them from spawning.

Added a few more names to the suffix list to ensure that the check won't have to roll as often.