Updated the name lists!
I took out some duplicate names and kept others as some of them are canon names and we keep cannon names in the normal pool until someone decides differently.

Didn't touch the names for wetlands and desert bc they're not implemented yet, there is no reason to limit names that cannot be gotten in game yet.