Edited 3+ group patrol chances
all non med patrols received an adjustment where 10 became the lowest base chance for 3+ group patrol success and 40 for the highest group patrol chance. Med patrols received 30 as their lowest, save for one type of patrol, and around 40 as their highest; as med patrols should be a little safer then ones with warriors in charge.

This should weight the outcomes on the cats abilities, giving more variety to outcomes.