Standardizing cat tags for scar and death text
- Since scar and death text is only shown in the cat's history screen and not on the patrol screen, r_c, m_c, p_l, s_c, and app1 are unnecessary here.

The tags being used for scar and death text were inconsistent, and patrol.py currently replaces any mention of r_c to the cat's actual name straight to the file, but this broke pronouns that used r_c and also made it so that the name wouldn't change if the player renamed the cat (for patrol events that didn't have p_l, app1, or m_c instead, that is).

Not sure what the desired behavior is (outputting the cat's name straight to the file or using the m_c tag?) but it should be easier to adjust one way or another now.