Fix attempt for p_l and r_c for patrol intro text
- Changed adjust_patrol_text to check if there is more than one cat on a patrol, rather than checking if there is a random cat, since the only time r_c should need to be anything else in this context is if there's only one cat on the patrol and they are p_l

- Fixed a crash in adjust_patrol_text

- Modified the check for possible_random_cats so that on a solo patrol, p_l = r_c, but on multi-cat patrols, p_l and r_c will not overlap.

- Minor spelling/grammar fixes in patrol.py

Sometimes, and it seems quite inconsistent, p_l and r_c will display with p_l's name for both in the intro text, but the tags will be correctly assigned once you hit "proceed". This occurs for patrols with proper tags for the intro text.

This is an attempt to fix the issue; I haven't run into it since.