added 2 patrols, updated an illness
Two new patrols added, one for a leaf-bare hunt and one for mountain border, both have a risk of death. Made carrionplace disease functional by adding healed text. Changed the values of the disease to be more deadly. Added infection herbs to help cure it, RBF is a bacterial infection.