Outside cats gain EXP over time
Someone in the dev-version-discussion channel on discord mentioned they thought it was weird that lost cats returned to the Clan with exactly the same EXP as they left with. I've had this code in my game for a while to give all outside cats EXP. It's a lot simpler than the apprentice EXP code gain, but it works.
The values I set are of course subject to change!