Old age death chance rework
- Changed old_age_death_chance from a linear equation to an exponential one
- Replaced old_age_death_chance with old_age_death_curve, an adjustable value from 1-7, in game_config.json

Turns out that the average life expectancy was less than *9 moons* after hitting the old_age_death_start threshold of 150 moons!! This spreads out the death chances a little more and allows cats to potentially live past moon 200 (though the chance is very low).

It also helps with senior cats that join from events. Previously, if a cat that joined from an event was near or over 200, they would die on the first timeskip in which they were healthy. Now they have a decent chance of living a few moons within the Clan.

I'll be sharing more of the math in a comment on my PR to explain the reasoning and balance behind this change.