Bug Fixing
Working on a few minor bugs and cosmetic issues, but were in the final stretch.

If your reading this as im working on it. idk. you might not check on it until I pull it out of draft mode and sent in a comment saying im ready for our next inspection XD