Redundant leader deaths
Changing death text that applies only to leaders and removing or changing them when necessary. For the most part, I changed deaths to apply to more ages/statuses or deleted them. In the cases of leader-specific deaths where the leader is saving a cat and dies in the process, I left those as leader-specific on the rationale that a leader might be more reckless with their lives.