Fix for cat list breaking bug
The issue would happen like this:

1. Open Clan list

2. Switch the view to another group (StarClan, COTC, etc) that has at least 1, but no more than 15, cat(s)

3. Click on one of these cats' sprites to open the profile view

4. Click the back button

Now all other lists of cats will be broken.

When using "back_button" to return to a list view with 15 or less cats, the cat_display container would be created with dimensions appropriate for the *smaller* number of cat buttons, meaning that if COTC had one row and your Clan had 4, all but the first row of your Clan will be cut off. Switching group views would not actually fix the cat_display dimensions because it would be modifying the preexisting cat_display. Now if it's less than 20 cats, the cat_display is deleted and a new, properly sized, one is generated. This might be inefficient but it's the only solution I could find that works reliably for now.