first update kittypet.json
changed "Cutter" to "vet"; kittypets know it as the vet iirc, and "Was taken to the Cutter recently" implies that, um. the Cutting occurred. which can be demonstrated as false if the kittypet joins the Clan later and has kits. also i believe that "vet" is used elsewhere in the thought code!